Update pgmspace include for SAMD and SAM
For both cores the pgmspace.h lives in the avr subdirectory. I don't know for which one the header lives in the root, however some (megaavr and mbed) do not have this file completely. But I don't simply want to remove it. May be better to define it as an empty macro if it doesn't exist yet.